refactor: modify utils to directory and index
